Fix NPE and iterator.hasNext() problems. Upgrade goby-io.jar to pre-Goby 1.9.5 version.
Goby 1.9.5 no longer supports queryLengths stored in the aligment header. This usage was deprecated in Goby 1.7. Versions of Goby 1.7-1.9.4 had a concat mode that transfered information from the header to the alignment entries transparently.
User will need to use the concatenate-alignments mode of Goby 1.7-1.94 when you need to migrate a 1.6- alignment to work with Goby 1.9.5+. The README in test/data/goby describes how the test alignment data were migration to work with 1.9.5. 


git-svn-id: https://igv.googlecode.com/svn/trunk@499 17392f64-ead8-4cea-ae29-09b3ab513800