Add command line plugin API.
A command line tool is defined by an XML file, which specifies the path to the tool (can be on system path),
the available commands, and the inputs. Right now the only allowable inputs are text (user types something in),
feature track (user picks a track and the contents are output to a file), and multi-feature-track.
A developer can write codecs for writing out / reading in features and distribute those as well,
should they so desire. IGV can only output in BED format right now, can read in any format.