Fix for JIRA-1884
This issue brought to light an issue with the TDF format outputted by igv tile. When tiling for the entire genome, the result is dependent on the specific ordering of chromosomes used. However, this ordering is never stored in the file. The previously used "chromosomes" attribute stored the chromosome names which were in the file being tiled, in the order in which they were found in the file. However, the whole genome view (properly) used the chromosome ordering provided by the Genome.  The result was that if the chromosome ordering was different when the TDF was loaded than when it was created, the whole genome display would be incorrect. Chromosome ordering was modified between 2.1.30 and 2.2.0 to put small chromosomes at the end, so files generated from <2.1.30 are likely to display incorrectly on 2.2.0 (and vice versa)

We take the following steps to correct this:

0. Write out the chromosomes in genome order to the TDF file, rather than file order.  As near as I can tell, the specific ordering of the chromosomes was not being used anywhere, so changing the outputted order should have no effect elsewhere.

1. When loading from TDFDataSource in whole genome view, call TDFDataSource.isChrAllValid, which examines the order of chromosomes in that field. If the field is absent, or if the ordering is different than the current Genome, the whole genome view is assumed to be invalid, and no data is returned

1a. If the TDF file is < 4 (current version as of this commit is 4, previous was 3), we know that the chromosome names are listed in file order but displayed in Genome order. We apply the old chromosome name comparator to the loaded chromosome names, and if they match the current genome we consider the data valid.

2. If the data is considered invalid, nothing is displayed for whole genome view. This is not ideal, might be better to recalculate or at least tell the user whats up (it gets logged but no dialog), but no data is better than wrong data.

3. TDF version has been changed to 4, and magic intro string is TDF4

Tests for the results of step 1 and 1a are included; we have wig files for dm3 and hg18, generated under 2.1.30 (version 3) and 2.2.1 (version 4)