Have GenomeManager test more intelligently for fasta file being absolute when loading .genome file
Previously, we only looked for a leading "/" or "\". The latter seems like it was intended to check against windows, except it didn't check for the presence of drive letters so it wouldn't have worked. We leave that check in but also use File.isAbsolute(). Write unit tests for loading genome where fasta is both relative and absolute; only really valuable to run on Windows.