Modularize GeneListManagerUI
We now are using the GeneListManagerUI dialog for 3 things:
0. Viewing a GeneList (original purpose)
1. Retrieving network from cBio ("Retrieve Network" button)
2. Exporting to TDM format

Previously, exporting to TDM was just another button. We now have GeneListManagerUI be created with text for the action button (default "View"), and a listener (GeneListListener) to receive a callback when that button is clicked. We have Gitools use this mechanism and open the GeneListManagerUI from an entry in the tools menu. The cBio entry could/should be migrated as well.