Modify build process to exclude certain jars from being unzipped
We were accidentally unzipping pretty much everything, now we exclude the libraries we want to exclude properly (batik-codec.jar, junit, appbundler, etc.). This was causing a problem with SVG export because batik needs information in the manifest. So the jars are included in the distribution as jars, rather than being unzipped and the contents copied