Add searching of negative strand to MotifFinderSource
Both the positive and negative strand are searched. What this means is that the sequence a user enters will be searched for on the  loaded sequence, and it's reverse complement. The feature is displayed at the positive-strand coords in either case.

This is not very efficient at the moment, we reverse-complement each chunk of sequence rather than the user-entered motif. This also means the returned order is backwards, so we need to do some reversing and combining of collections to give features in the proper order. Since this developer doesn't know how to reverse-complement a regex, we went this way.

Refactored some test code as well

IGV-1905 #resolve