Modify plugin decoders to read header but remember lines
We had this problem with tribble codecs before, because one never knows how long a header is, typically they just keep reading until they hit lines which no longer look like headers. This often means missing the first data line. This was solved by opening a separate stream on the file. We can't do that for plugins which read from stdout, so we create a custom LineReader which stores each line once it's been read for later retrieval. Not sure how useful this is generally but it should work for this purpose.