Handle the cancelling/failure of sequence better
We would update the stored genome list file even if the sequence were cancelled, giving the false impression it was a local genome. This resolves the issue that if a genome is downloaded, with it's sequence, but the sequence is cancelled, it would still appear in the "Load Genome From Server" list.

However, a fully downloaded genome still appears in that list, which may be confusing for users.