Increase scroll amount when left/right key pressed
The 'left' and 'right' keys are not really helpful right now since they scroll too little. I'd like to have a way to 'scan' through a reference just by keeping a button pressed. With this change, the left and right keys scroll by 50 pixels instead of 5, which improves usability, in my opinion.