Show base mismatches at lower zoom
Show base mismatches at lower zoom levels to enable visualization
of haplotype phase in long range data like PacBio and 10X.  To phase
variants, it is important to see which derive from the same molecule.
That is not easy to visualize unless variants are shown at low zoom
levels that include tens of thousands of bases.