Trigger refresh when large insertion settings change
Trigger a new render of the alignment tracks when the configuration
settings that define a "large insertion" are changed.

Also, update the options used to render an alignment track when the track
is sent an explicit "reload" notification.  This is important, as
preferences that affect rendering could have changed since the reload
options were initialized.