Label the size of large indels
Transform the "flag large insertions" option into an option to label
the size of large insertions and deletions.  For large events, the
size of the event (in basepairs) is show in a label.  This is particularly
useful to differentiate insertion events of different scales and to
see the exact basepair size of deletions.