Highlight clipped alignment edges
Read clipping often indicates events of interest, including assembly
errors and structural variants.  IGV users often activate the option
to "show soft-clipped bases" to see clipping, but that view is difficult
when zoomed-out and can be overwhelming with long reads that can have
thousands of basepairs clipped.

As an alternative, add an option to highlight/flag clipping of more
than a specified number of basepairs.  This draws a pink border on
any edge that has enough basepairs clipped.