Bugfix: Make alignment grouping consistent and persistent
This fixes a bug in the behavior of the "group alignments by" feature.
Previously, when a grouping option was selected, it was applied to all
loaded BAM tracks.  However, when a new BAM track was loaded, no grouping
was applied.  Further, when the data panel was re-rendered, such as when
a filter option (e.g. "Filter duplicated reads") is changed, the grouping
was lost.  This adds sharing and persistence of the state of the "group
alignments by" option.