Drop unnecessary logic
Drop unnecessary grouping logic.  When the "group alignments by" setting
is changed for a track, the track communicates that up to the top-level
controller, which then updates all tracks.  It is not necessary for the
track to update itself before making the top-level request, since the
controller will simply updated it again.