Create graphics for drawing strand indicator lines
In "squished" mode, there is not sufficient vertical space to draw
an arrow to indicate read strand; so, a dark gray line at the end
of the read is used as an alternative.  Before this fix, the same
graphics was used to draw this line as to render the alignment blocks.
As a result, some blocks were incorrectly drawn as dark gray. This
avoids that problem and improves clarity by creating a separate
graphics context for drawing the strand indicator lines.