Only list suppressed tags as hidden in alignment popup
The alignment details popup message has a section labeled "hidden
tags" that lists the SAM tags that are set in preferences to be
hidden.  Modify that message to only list tags that are suppressed
in preferences and actually present in the read being viewed.

For example, if "MD,SA" are set to be hidden in preferences but an
alignment on has "MD" and not "SA", then only list "MD".