JavaFX Cytoband header, plus clean-up
Improved the ported Cytoband header code, to the point where it's
displaying what looks like an accurate image.  It still needs work, in
that it is somewhat fuzzy compared to the Swing version.
Wired the Cytoband header to the ReferenceFrame via a JavaFX Bound
Property, also wired to the ChromosomeComboBox.
Simplified the ChromosomeComboBox based on this wiring.
Inlined the CytobandRenderer, as we expect there will never be any need
for an alternate implementation.
Removed a number of copied versions of "legacy" classes (IGV,
GenomeManager, etc), going back to the originals (where needed) or
avoiding altogether.  Added a "backend placeholder" for any custom code
of the sort embedded in e.g. the IGV class.
All around better handling of component sizing via binding.
Pushed the row ScrollPane wrappers down into IGVRow (and subclasses) as
they have minimal use by any external component.
Fixed an issue with the ResizableCanvas diomensions being bound to the
wrong container Pane properties.