JavaFX: interim check-in for handling multiple ReferenceFrames
Experiment with managing width for multiple HeaderPanes using Bound
Properties.  This fixes the issues with rendering on resize but is not
final.  The Property should be better located where the rest of the UI
can use it (i.e. Tracks), and other related values should be bound to
it as computed properties.
Cleans up some debug logging.