JavaFX: wire up H/W props for Track rendering components
Wires the height & width properties of DataPane & Container and sets
initial height.  This is just a start, but probably a better approach
is to push this type of stuff into a common superclass for all the
rendering classes.
Fix to GenomeManager, allowing start-up underJavaFX without requiring
init of IGV singleton & main java.awt.Frame.