Updating project build and layout for Java 9+
Modified the Gradle build for Java 9+.  Right now it's focused on Java
9 instead of being switchable between 9 & 8.
Separating libraries into classpath JARs (in 'lib') and module path
JARs (in 'modules').  Likewise for test JARs.  In the end, we probably
want all CP JARs moved to MP (or as many as possible) but we're not
there yet.
Removed several libraries that seem to be unused (may have been Batik
dependencies).  Unpacked the jlfgr JAR into a dedicated resource
directory as it was too difficult to deal with as a "resource-only" JAR.
Stripped extra classes out of the goby & na12878 libs as these were
causing split packages.  May have more to do in that regard.
Modified the Java 8 check in the start-up to guard against Java 8
instead of require it.