New Windows launcher bat to avoid path issues
Fixes an issue on Windows 10 where the Windows 8.3 "short name" may
not exist for all path components, particularly on drives other than
C:, allowing spaces into the path (e.g. "Program Files").  The new
launcher script uses relative paths and relies on the shortcut to
start it in the correct directory.
We keep the old script for existing users with batch jobs.