Builds new bundles, both with and without embedded JRE.
New bundles for all platforms, with and without embedded JRE.  We're
aiming for the "no Java" bundles to eventually be the default and so
these get the simpler bundle names.
Also, modified the launchers to pick up an (optional) java_arguments
file from the user's home directory (in a '.igv' folder).  Can be used
for custom java args such as -Xmx memory spec.
The file & folder will be created if missing.