Fixed the Mac app launcher
This deals with a regression formerly fixed a while ago.  The Mac app
launcher should *not* pass along extra command-line arguments (unlike
the igv.sh and igv.command launchers).  When launched as a Mac app,
some versions of macOS will pass extra parameters through to the script
that are nonsensical to IGV.  Users with needs for that should use
those other scripts instead.