Changed ToolTip text for better Windows high-contrast support
For some reason, this widget does not have its foreground (text) color
appropriately set in high-contrast mode on Windows, and potentially
other situations as well.
The fix is to pick up the proper color from a widget that *does* change
appropriately, then wrap the text content in a <div> to set that color
explicitly.