Improve mac packaging (#1012)
* Improved Mac App launcher for use with shc

Improved the Mac App launcher script so that it would work better with
the 'shc' Bash script compiler.  Notes: make Bash be the expected shell
explicitly; simplify the conditional for dealing with the java_arguments
file.

* Added a comment to the Mac App launcher script

This gives compilation instructions.
Also, renamed the script.

* Added a compiled binary for the Mac App launcher

This is the binary executable produced by compiling IGV.sh with 'shc'.
Instructions are in IGV.sh.  This only needs to be recompiled if that
script changes and so it's not included as part of the Gradle script.
Updated the Gradle script to exclude IGV.sh from the bundle.