Update with many changes to improve supplemental alignment display and navigation.
Major rework of the previous alignment dialog popup to make it interactive and clearer.
Use it to navigate between reads in a supplemenary alignment group.
   Click to move to the selected read.
   Shift + Click to add it as a split screen.

When jumping to a selected read/mate the selected reads will be now be sorted to the top.

Renamed several classes and moved them to the ui/supdiagram package
Refactoring some methods around Frame creation, this could still be improved.