Make delete/backspace delete tracks in dev/prod mode
* There is a wierd issue where the delete key removes selected tracks but only if IGV is in develop mode.
However, there's a different weird issue where the production IGV seems to be set to develop mode.
Since it seems like this defaults to being on in practice it's likely that whatever worry made this a developent mode
option is probably unfounded.

This removes the isDevelopment() guard and does some minor refatoring in the same class.