Fix bug in zoom calculation. (#1426)
Fixing a rounding error  in calculateZoom.  Instead of rounding to the closest integer zoom level I believe
we should be rounding up to the nearest zoom level which can contain the given window.

This fixes a display issue with the ZoomSliderPanel.  I'm not 100% sure I understand the potential consequences
of changing thhis for other code that relies on it, so it would be good to have a second pair of eyes on it.
It seems intuitively to be correct now but it's possible the use case of "zoom level which is closest to the width of the given window"
was really what was wanted.  I haven't found any obvious issues with some manual testing of this change.