Fix null pointer exception in SAMAlignment.getAlignmentValueString (#1772)
Commit `eecf30a591` introduced a code path that checks if a particular
render option is set.  But sometimes this function is called with `null`
for the `renderOptions` argument (e.g. in `getClipboardString`), which
results in a null pointer exception, like when using "Copy read details"
in the right click menu.

This adds a null check to avoid the NPE so we can copy read details
again.